requirements.txt
amqp==5.1.1
ansicon==1.89.0
APScheduler==3.6.3
arrow==1.2.2
asgiref==3.5.2
attrs==21.4.0
backports.zoneinfo==0.2.1
bcrypt==3.2.2
beautifulsoup4==4.11.1
billiard==3.6.4.0
blessed==1.19.1
cachetools==4.2.2
celery==5.2.6
certifi==2022.5.18
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
cryptography==37.0.2
distro==1.7.0
Django==3.0.0
django-picklefield==3.0.1
django-q==1.3.9
django-redis==5.2.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
fastnumbers==3.2.1
flower==1.0.0
gevent==21.12.0
google-api-core==2.8.0
google-api-python-client==2.49.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
googleapis-common-protos==1.56.1
greenlet==1.1.2
gunicorn==20.1.0
httplib2==0.20.4
humanize==4.1.0
idna==3.3
jinxed==1.2.0
jsonschema==3.2.0
kombu==5.2.4
lxml==4.8.0
oauth2client==4.1.3
oauthlib==3.2.0
paramiko==2.11.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.20.1
psycopg2-binary==2.8.6
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyNaCl==1.5.0
pyparsing==3.0.9
Pyrogram==2.0.24
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-telegram-bot==13.11
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.4.1
redis==3.5.3
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
texttable==1.6.4
tornado==6.1
tzdata==2022.1
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
websocket-client==0.59.0
zope.event==4.5.0
zope.interface==5.4.0
